// src/api/generated/models/gymSignupRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * GymFlow API
 * OpenAPI spec version: v1
 */

export interface GymSignupRequest {
  gymName?: string;
  address?: string;
  city?: string;
  country?: string;
  phone?: string;
  gymEmail?: string;
  adminName?: string;
  adminEmail?: string;
  password?: string;
}